Move the case_change_service to python_utils
