+++ core/controllers/classroom_test.py
@@ -23,7 +23,7 @@ from core.domain import topic_domain
 from core.domain import topic_services
 from core.tests import test_utils
 import feconf
-from scripts import case_change_service
+import python_utils
 
 
 class BaseClassroomControllerTests(test_utils.GenericTestBase):
@@ -98,7 +98,7 @@ class ClassroomDataHandlerTests(BaseClassroomControllerTests):
                 topic_services.get_topic_summary_by_id(topic_id_2).to_dict())
 
             expected_dict = {
-                'topicSummaryDicts': [case_change_service.camelize(
+                'topicSummaryDicts': [python_utils.camelize(
                     topic_summary_dict)]
             }
             self.assertDictContainsSubset(expected_dict, json_response)
